Switched up application structure once again, should be the final structure. Much cleaner now. Each expression type gets its own entry_definitions module in seperate files which will contain all validation and links.
