--- junto-rust/zomes/core/code/src/expressions/definitions/expression_validation/mod.rs
@@ -1,5 +1,0 @@
-pub mod channel;
-pub mod group;
-pub mod post;
-pub mod time;
-pub mod user;
