+++ junto-rust/zomes/core/code/src/expressions/definitions/mod.rs
@@ -1,4 +1,6 @@
-pub mod entry_definitions;
 pub mod app_definitions;
-pub mod expression_validation;
-pub mod expression_links;
+pub mod channel_entry_definitions;
+pub mod group_entry_definitions;
+pub mod post_entry_definitions;
+pub mod time_entry_definitions;
+pub mod user_entry_definitions;
