+++ junto-rust/zomes/core/code/src/lib.rs
@@ -22,12 +22,12 @@ mod expressions;
 
 define_zome! {
     entries: [
-       expressions::definitions::entry_definitions::user_definition(),
-       expressions::definitions::entry_definitions::time_definiton(),
-       expressions::definitions::entry_definitions::channel_definition(),
-       expressions::definitions::entry_definitions::group_definition(),
-       expressions::definitions::entry_definitions::post_definition(),
-       expressions::definitions::entry_definitions::resonation_definition()
+       expressions::definitions::user_entry_definitions::user_definition(),
+       expressions::definitions::time_entry_definitions::time_definiton(),
+       expressions::definitions::channel_entry_definitions::channel_definition(),
+       expressions::definitions::group_entry_definitions::group_definition(),
+       expressions::definitions::post_entry_definitions::post_definition(),
+       expressions::definitions::post_entry_definitions::resonation_definition()
     ]
 
     genesis: || {
